requirements.txt
black
flake8-black
mobase-stubs
PyQt6